Split Chart1 out of axis-chart
Most of axis-chart.js is moved to chart1.js, so I did a rename to chart1.js in the previous commit and this commit moves the remainder back to axis-chart.js, i.e. axis-chart component and its attributes : blockService, className, didRender, blockFeatures(), featuresCounts(), drawBlockFeatures0(), drawBlockFeaturesCounts(), drawBlockFeatures(), redraw(), parseTextData(), pasteProcess().  layoutAndDrawChart() (containing Chart1) remains; next step will move layoutAndDrawChart into class Chart1.
layoutAndDrawChart() : change some this to axisChart.
chart1.js: import path ../utils/ is now ../.

Handle extra data in Feature.value, after the range .value[0..1], e.g. .value[2] :
block-features.js blockFeatureLimits() : group : slice value to get [0..1] and filter out null values.
paths-aggr.js : valueBound() : address value[1] instead of value[length-1] - same effect, ignores extra data.

backend/common/utilities/ :
 30787d0  4839 Jan  5 21:19  block-features.js
 74c8a7c 28697 Jan  5 21:53  paths-aggr.js
frontend/app/ :
 06dd3ad 13950 Jan  5 15:55  utils/draw/chart1.js
 ea5ac9f  7888 Jan  5 22:41  components/axis-chart.js
